Add replay parsing code Update replay labeling code Update IDM training script Move stuff around
